requirements.txt
discord.py==2.4.0
aiohttp==3.9.5
Flask==3.0.0
